requirements.txt
bledom==0.0.9
python-dotenv>=1.0.1
pillow>=10.2.0
mss>=9.0.1
tweener>=1.0.4